refactor(evaluator): part the value rebuild from the dispatcher
`evaluate/mod.rs` changed for three reasons: a node handler added to the
dispatcher, a value shape added to the rebuild, and a test module
registered. The first two are apart now -- the rebuild is
`evaluate/rebuild.rs`, and the file it left is 200 lines shorter.

Nothing moves for a caller: the module re-exports what it held, so
`evaluate::evaluate_result_vec_to_array_expr` is the path it always was.
The three test files that read a factory through the dispatcher's imports
now import it themselves.
